Move remaining cache objects to service pkg
